Delay pytorch imports as much as possible
* Move some torchdata code from psp/dataset.py to psp/training.py. We
  moved what is needed in training only so that configs can still import
  psp/dataset without required pytorch.
* Add TYPE_CHECKING checks in a few places where it speeds things up
* Delay pytorch and psp/training.py (which needs to import pytorch for
  some class definitions) imports. This speeds up calling train_model.py
  and eval_model.py by 2-3 seconds when calling them  with "--help".
